// src/app/page/master/map-angka-to-bulan/map-angka-to-bulan.interface.ts
import { EventEmitter } from '@angular/core';

export interface MapAngkaToBulan {
	kdProduk?: string;
	kdEvent?: number;
	label?: any;
	value?: any;
	kdMapping?: any;
	kode?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}